EntiTrack_API/requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
flask-restx
google-genai
flask-cors
langchain
langchain-google-genai
pydantic
pandas
scikit-learn
setuptools
wheel
spacy